Extract methodology into its own app
